Se organizan y agregan funciones en clase dbutil. Agrega DELETE estimate, repara POST 23_05_2023 10_35AM
