+++ Infrastructure/IEstimateDetailDBRepository.cs
@@ -5,4 +5,6 @@ namespace WebApiSample.Infrastructure;
 public interface IEstimateDetailDBRepository : IGenericRepository<EstimateDetailDB>
 {
      Task<IEnumerable<EstimateDetailDB>>GetAllByIdEstHeadersync(int code);
+      public Task<int> DeleteByIdEstHeaderAsync(int IdEstHeader);
+     
 }
